Extract shared HydroShare card helpers into reusable module
- Move reusable functions out of HydroShareResourcesCards
- Enable reuse across HydroShare card components
